Refactored the core event system to align with SOLID principles.
- **Decoupled `Event` from `EventManager`:** Removed the static `Event.Send()` method and its reliance on a `ServiceLocator`. Events are now simple data containers.
- **Introduced `IEventBus`:** Created a new `IEventBus` interface and an `InMemoryEventBus` implementation to handle event dispatching.
- **Dependency Injection:** The `IEventBus` is now injected into services and character instances, promoting loose coupling and testability.
- **Updated Call Sites:** Replaced all `event.Send()` calls with `character.EventManager.SendEvent(event)`, ensuring the entire system uses the new event bus.
- **Improved Test Quality:** Fixed all test failures resulting from the refactor and improved numerous tests by adding proper mocking and assertions, resolving race conditions, and correcting invalid test setups.

+++ Hagalaz.Game.Scripts.Tests/Commands/DrawModelCommandTests.cs
@@ -17,7 +17,7 @@ namespace Hagalaz.Game.Scripts.Tests.Commands
             // Arrange
             var widgetMock = Substitute.For<IWidget>();
             var widgetsMock = Substitute.For<IWidgetContainer>();
-            widgetsMock.GetOpenWidget(Arg.Any<short>()).Returns(widgetMock);
+            widgetsMock.GetOpenWidget(Arg.Any<int>()).Returns(widgetMock);
             var characterMock = Substitute.For<ICharacter>();
             characterMock.Widgets.Returns(widgetsMock);
